FastAPI/requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
databases
python-dotenv
web3
pymysql
aiomysql
ipfshttpclient
